requirements.txt
django
django-filter
djangorestframework
djangorestframework-simplejwt
httpie
sqlite
